Lets add a `screenBackgroundColor` in navigatorStyle
(c) https://github.com/MattDavies/react-native-controllers/commit/fb5dec7d4e53026cd11617ac86b093e4397c0df0

By default the view backgroundColor is white. This produces a white flickering for some of us when using dark backgroundColor in the <View />, because when we use push a new RootView is created and React styles applies not directly. So what happens is the following:

0. Navigator takes the registered component
1. Create a new ReactView, which has by default backgroundColor white!
2. Applies the navigation
<Here we see the flickering>
3. React styles will get applied

This commit adds a new navigatorStyle property `screenBackgroundColor`, which will get applied directly to the view as the backgroundColor, before React applies the backgroundColor from the actual styles.